libopendkim: fix dkim_sig_keybits() for ed25519 keys
In dkim_sig_keybits() the wrong size was being returned for ed25519.
For now, always return 256 since ed25519 is always 256-bit.  The
underlying bug is in the calculation of 'keybits' in various other
dkim functions and that will be fixed in a subesquent PR.

* libopendki/dkim.c (dkim_sig_keybits): as above

(This is one of split portions of commit
e49e90f0fa342552eab5cf6335cde9728aa6e502 in jcastle-gh/ed25519-tests branch)